requirements.txt
flask
flask-smorest
python-dotenv
flask-jwt-extended